EditarAlquiler01

+++ src/main/java/com/FoodSpringApp/FoodSpringApp/repository/AlquilerRepository.java
@@ -8,6 +8,6 @@ import org.springframework.stereotype.Repository;
 import com.FoodSpringApp.FoodSpringApp.model.Alquiler;
 
 @Repository
-public interface AlquilerRepository extends JpaRepository<Alquiler, Long> {
+public interface AlquilerRepository extends JpaRepository<Alquiler, Integer> {
     List<Alquiler> findByClienteId(int clienteId);
 }
